[HOUSEKEEPING] refactor spk setup command
